✨ feat: 完成数据加载存储

+++ Notes/ref/Ini.h
@@ -54,7 +54,7 @@ public:
 	 * @{
 	 */
 	bool Read(CString sSection,   CString sKeyName,   CString& sValue);
-	bool Read(CString sSection,   CString sKeyName,   DWORD&   dwValue);
+	bool Read(CString sSection,   CString sKeyName,	  ULONG&   dwValue);
 	bool Read(CString sSection,   CString sKeyName,   int&     nValue);
 	bool Read(CString sSection,   CString sKeyName,   UINT&    nValue);
 	bool Read(CString sSection,   CString sKeyName,   long&    nValue);
